Allow calling functions that may not exist
This crate uses Cargo features to define the API, gating access to APIs
which weren't enabled. This isn't necessarily always the desired use
case, however. Sometimes a binary wants to be compiled as maximally
flexible (e.g runs almost anywhere) but still call newer APIs where it
can.

To accommodate this use case, this commit implements a fallback for APIs
where features are not enabled. Currently this is just a
proof-of-concept for one function to gauge interest, but the theory is
that this could generally apply to many APIs to provide a uniform API
for this crate across all OS versions.

Technically speaking, this enables a feature where when an FFI function
wouldn't otherwise be defined we define our own. Our own wrapper then
looks up the symbol at runtime (via `dlopen` + `dlsym`) and caches the
result in a global. The wrapper then also provides a fallback
implementation which typically just returns an error. This is similar to
what's done in the standard library as well.