Add ClientBuilder::handshake2 for async clients
Currently the ClientBuilder::handshake function doesn't work with asynchronous
clients as it just does a loop internally. This function extracts that logic
into a new `MidHandshakeClientBuilder` type and `ClientHandshakeError` enum to
handle the loop externally.

This should allow asynchronous clients to still use `ClientBuilder` and the
`anchor_certificates` function, namely allowing `native-tls` to update to using
`ClientBuilder` and leveraging this logic.