ClientBuilder::add_anchor_certificate
`anchor_certificates` is not flexible enough: it overrides previously
set certificate.  In my wrapper layer there's a function
`add_anchor_certificate`, and it's not possible to imlpement in
current API.

Alternatively we can change `anchor_certificates` to extend the
certificate list, not replace, but it will be technically
semver-incompatible change.