New Item search features
This commit adds two new features to the Item search:

* Support for kSecMatchLimitAll. Previously, this library only allowed
searching for a single item, or for a specific number of items. This
feature allows all items available to be returned.

* Support for kSecAttrAccessGroup, specifically set to
kSecAttrAccessGroupToken. This allows retrieving items that have been
retrieved from a token, such as a smart card, via a CryptoTokenKit
provider.