First pass at password API support on iOS.
The current API support for passwords in keychain uses a legacy API (including support for creating file-based keychains) which runs only on macOS.  But the base API for Keychain Items runs on both iOS and MacOS, and can do everything that the legacy API does in the default keychain.

This change adds a `passwords` module to the base security-framework wrapper API that can be used on both iOS and macOS.  This provides entries for setting, getting, and deleting generic passwords.

 This change also alters/extends the existing `item` module as follows:
- The iOS and macOS search result interpretation have been merged, so iOS search can return all types of items.
- Constants for generic password attributes have been added to the underlying `-sys` framework.
- Updated formatting by `rustfmt` has been applied.

Finally, this change adds a crate with a skeletal ios-shim library and tests for interoperability between the macOS and iOS APIs for handling generic passwords.  An actual iOS app and support for internet passwords will be added in a future commit.