Update the shim to use CF types in its interface.
There's no point in using a straight C interface for an iOS layer, because it makes memory management unnecessarily difficult.  By using CoreFoundation types in the interface, we can return allocated memory to our callers when returning passwords.

This also means we don't need to add anything to the underlying sys interface for the shim.