Fix iOS compilation errors and warnings.
1. iOS has had `SecTrustEvaluateWithError` since iOS 12, so make sure `evaluate_with_error` is defined and used on iOS.
2. Conditionalize `use` statements so they match the compilation platform.

NOTE: There are no features for older versions of iOS, unlike those for MacOS.  So this crate may not compile for versions of iOS prior to 12. But when you consider that Apple doesn't allow apps that haven't been recompiled for iOS 12 or later, it's not clear whether such features are warranted.