Make SecTrust::evaluate_with_error use dlsym when !OSX_10_14
When `feature = "OSX_10_14"` is specified, we use the symbol
`SecTrustEvaluateWithError` directly. When the feature is disabled, we
first try to look it up using `dlsym`, and if the symbol doesn't exist,
only then do we fall back to `SecTrustEvaluate`. This gives nicer error
messages for projects that use `SecTrust::evaluate_with_error`, while
still supporting macOS 10.12.